Initial support for persistent reminders
The architectural difficulty here is that we might need to "respond" to messages
after a CoffeeBot restart. Reminder orders are persisted to the DB together with
the channel or user they need to be sent to, so we need to be able to send these
messages without a regular `Handle` at hand. Hence the multi-channel handler.

While researching the way to send messages to a random channel, I've also come
across the concept of Snowflake, which is how Discord keeps track of users / channels
even after they have been renamed. The reminders code stores Snowflakes instead of
names for users / channels. This has required passing this through to Valid messages,
and unfortunately isn't really compatible with the existing message log feature
(not a huge deal because it's only used to recompute lisp state).